+++ publish/CHANGELOG.md
@@ -2,6 +2,9 @@
   - Adds a new field `clearArea` to the `expand_vegetation.yaml`.
   - Adds a new value `all` to the `randomDamage` field of `expand_locations.yaml` to affect all pieces.
   - Changes the `randomDamage` field of `expand_locations.yaml` to not affect pieces with custom health.
+  - Fixes location reset from Upgrade World mod creating duplicate terrain compilers when terrain leveling was used.
+  - Fixes terrain leveling using the modified terrain height, which caused wrong terrain height when Upgrade World was used.
+  - Improves performance of terrain leveling.
 
 - v1.18
   - Fixes locations sometimes not appearing until relog.